# Use more efficient AttributeParentConnectingVisitor