1:1 copy of php-parsers ParentConnectingVisitor but optimized for our use-case

after PR
```
➜  php-code-coverage git:(attr-paren) ✗ hyperfine --prepare 'rm -rf /tmp/coverage-cache' 'php test.php'
Benchmark 1: php test.php
  Time (mean ± σ):     159.4 ms ±   1.2 ms    [User: 137.7 ms, System: 18.9 ms]
  Range (min … max):   156.2 ms … 161.1 ms    18 runs
```

before PR
``` 
➜  php-code-coverage git:(attr-paren) ✗ hyperfine --prepare 'rm -rf /tmp/coverage-cache' 'php test.php'
Benchmark 1: php test.php
  Time (mean ± σ):     163.0 ms ±   1.1 ms    [User: 140.1 ms, System: 19.9 ms]
  Range (min … max):   161.1 ms … 164.8 ms    17 runs
```

refs https://github.com/sebastianbergmann/php-code-coverage/issues/1072#issuecomment-2832263578

Less references between AST nodes mean less work for GC